requirements.txt
selenium==3.141.0
requests==2.23.0
schedule==0.6.0
PyYAML==5.3.1
